Move also domain models and parent classes
