+++ decisiontable-quarkus-example/src/test/java/org/kie/kogito/decisiontable/quarkus/NativeRestQueryIT.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2020 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,15 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.kogito.examples;
+package org.kie.kogito.decisiontable.quarkus;
 
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
+import io.quarkus.test.junit.NativeImageTest;
 
-@SpringBootApplication(scanBasePackages = { "org.kie.kogito.**", "http**" })
-public class KogitoSpringbootApplication {
+@NativeImageTest
+public class NativeRestQueryIT extends RestQueryTest {
 
-    public static void main(String[] args) {
-        SpringApplication.run(KogitoSpringbootApplication.class, args);
-    }
+    // Execute the same tests but in native mode.
 }
