+++ process-optaplanner-quarkus/src/test/java/org/kie/kogito/examples/quarkus/DefaultWorkItem.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.kogito.examples;
+package org.kie.kogito.examples.quarkus;
 
 import java.util.Map;
