+++ process-optaplanner-springboot/src/main/java/org/kie/kogito/examples/springboot/FlightSeatingApplication.java
@@ -13,15 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.dmn.kogito.springboot.example;
+package org.kie.kogito.examples.springboot;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-@SpringBootApplication(scanBasePackages = { "org.kie.dmn.kogito.**", "org.kie.kogito.**", "http**" })
-public class KogitoSpringbootApplication {
+@SpringBootApplication(scanBasePackages = { "org.kie.kogito.**" })
+public class FlightSeatingApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(KogitoSpringbootApplication.class, args);
+        // Spring-boot dev tools class loader cannot detect generated sources
+        // so we need to disable restart if we want to take advantage of live-reload
+        System.setProperty("spring.devtools.restart.enabled", "false");
+        SpringApplication.run(FlightSeatingApplication.class, args);
     }
 }
