+++ process-service-calls-quarkus/src/main/java/org/acme/travels/quarkus/Traveller.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.acme.travels;
+package org.acme.travels.quarkus;
 
 public class Traveller {
