+++ process-service-calls-springboot/src/main/java/org/acme/travels/springboot/Address.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.acme.travels;
+package org.acme.travels.springboot;
 
 public class Address {
