+++ process-service-calls-springboot/src/main/java/org/acme/travels/springboot/Traveller.java
@@ -13,33 +13,27 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.acme.travels;
+package org.acme.travels.springboot;
 
-public class User {
+public class Traveller {
 
-    private long id;
-    private String username;
     private String firstName;
     private String lastName;
     private String email;
-    private String password;
-    private String phone;
-    private int userStatus;
+    private String nationality;
+    private Address address;
 
-    public long getId() {
-        return id;
+    public Traveller() {
+
     }
 
-    public void setId(long id) {
-        this.id = id;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
+    public Traveller(String firstName, String lastName, String email, String nationality, Address address) {
+        super();
+        this.firstName = firstName;
+        this.lastName = lastName;
+        this.email = email;
+        this.nationality = nationality;
+        this.address = address;
     }
 
     public String getFirstName() {
@@ -66,27 +60,26 @@ public class User {
         this.email = email;
     }
 
-    public String getPassword() {
-        return password;
+    public String getNationality() {
+        return nationality;
     }
 
-    public void setPassword(String password) {
-        this.password = password;
+    public void setNationality(String nationality) {
+        this.nationality = nationality;
     }
 
-    public String getPhone() {
-        return phone;
+    public Address getAddress() {
+        return address;
     }
 
-    public void setPhone(String phone) {
-        this.phone = phone;
+    public void setAddress(Address address) {
+        this.address = address;
     }
 
-    public int getUserStatus() {
-        return userStatus;
+    @Override
+    public String toString() {
+        return "Traveller [firstName=" + firstName + ", lastName=" + lastName + ", email=" + email + ", nationality="
+                + nationality + ", address=" + address + "]";
     }
 
-    public void setUserStatus(int userStatus) {
-        this.userStatus = userStatus;
-    }
 }
