+++ process-service-rest-call-quarkus/src/main/java/org/acme/travels/services/AuditService.java
@@ -17,7 +17,7 @@ package org.acme.travels.services;
 
 import javax.enterprise.context.ApplicationScoped;
 
-import org.acme.travels.User;
+import org.acme.travels.quarkus.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
