+++ process-service-rest-call-quarkus/src/main/java/org/acme/travels/services/UserService.java
@@ -18,7 +18,7 @@ package org.acme.travels.services;
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 
-import org.acme.travels.User;
+import org.acme.travels.quarkus.User;
 import org.acme.travels.rest.UsersRemoteService;
 import org.eclipse.microprofile.faulttolerance.Fallback;
 import org.eclipse.microprofile.rest.client.inject.RestClient;
