+++ process-service-rest-call-springboot/src/main/java/org/acme/travels/services/AuditService.java
@@ -15,7 +15,7 @@
  */
 package org.acme.travels.services;
 
-import org.acme.travels.User;
+import org.acme.travels.springboot.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
