+++ serverless-workflow-functions-quarkus/src/test/java/org/acme/numbers/serverless/workflow/functions/RestExampleTest.java
@@ -17,6 +17,7 @@ package org.acme.numbers.serverless.workflow.functions;
 
 import java.util.Collections;
 
+import org.acme.numbers.NumbersMockService;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
